исправил все предупреждения, добавил врага, сделал смерть врага

+++ source/Animation.hpp
@@ -139,9 +139,9 @@ public:
 
 	bool isPlaying()  {return animList[currentAnim].isPlaying;}
 
-	float getH()  {return animList[currentAnim].frames[0].height;}
+	int getH()  {return animList[currentAnim].frames[0].height;}
 
-	float getW() {return animList[currentAnim].frames[0].width;}
+	int getW() {return animList[currentAnim].frames[0].width;}
 	//////////////////////////////////////
 };
